musvg: fix subtle bug introduced by moving strings into attr_storage
problem exists for writes of a string id into an attribute that
allocates a string, because the string allocation can cause the
attribute to move, and if the client has a pointer to the attribute
it may no longer be valid by the time the string id is written.

fix is to revalidate attr pointer after alloc_string.

separate string table would improve the usability. the routines
only write one attribute at a time, so it would be sufficient to
move strings into their own table.